Replace the device #ifdef debug gates with log categories; fix PTE shifts
All compile-time debug gates in the device and machine code are now
runtime log calls (house rule 9.7), each compiled in every build:

- floppy: DEBUG_FLOPPY_DMA, DEBUG_FLOPPY_PIO and FLOPPY_DIAG at debug,
  DEBUG_DETAIL and DEBUG_DATATRANSFER_PIO at trace (30 blocks).
- hdlc: HDLC_DEBUG and DMA_DEBUG at debug, DEBUG_DETAIL and
  RX_BLAST_LOGGING at trace. HDLC_LOG()/HDLC_LOG_IRQ() are LOG() calls;
  the four register decoders are one table-driven hdlc_log_bits() that
  writes one line instead of a line in pieces. DMAEngine_Log() and
  DMAControlBlocks_Log() go through the logger instead of printf into
  stdout (the guest terminal on native builds).
- rtc (DEBUG_RTC), term (DEBUG_TERMINAL), device (LOG_DEVICE_NOT_FOUND,
  now debug level).
- The commented-out "#define DEBUG_x" lines are gone.

Compiling the old gated code showed what it had rotted into: a DMA debug
message referred to a variable (disp) that the step 2 cleanup had
removed as unused, because its only reader was compiled out; the label is
now computed where it is printed. One RTC gate held an "else" of an outer
if; that is now plain code after the if.

A boot in the ASan/UBSan build with --log='*:trace' (every category on)
found undefined behaviour in production code: cpu_mms.c shifted a 16-bit
page-table word left by 16 as an int (ConvertFrom16BitPTE,
GetPageTableEntry twice, and the PT_Write trace), which overflows when bit
15 (WPM) is set. The operand is widened to uint32_t first. With that, a
20M-instruction boot with every category at trace runs clean under the
sanitizers.

Release and Debug native and glass WASM builds have no warnings from
project sources; 17/17 tests pass in Release, Debug and the sanitizer
build.

+++ src/cpu/cpu_mms.c
@@ -77,7 +77,7 @@ bool CreatePagingTables(void)
 // Helper functions
 static uint ConvertFrom16BitPTE(ushort value)
 {
-    uint pte = (uint)((value & 0xFE00) << 16 | (value & 0x01FF));
+    uint pte = ((uint32_t)(value & 0xFE00) << 16) | (uint32_t)(value & 0x01FF);
     return pte;
 }
 
@@ -185,12 +185,12 @@ void PT_Write(uint address, ushort value)
             if ((address & 0x01) == 0)
             {
                 // Even address
-                pageTableEntry = (uint)(value << 16 | g_paging_tables.shadowRam[offset + 1]);
+                pageTableEntry = ((uint32_t)value << 16) | g_paging_tables.shadowRam[offset + 1];
             }
             else
             {
                 // Odd address
-                pageTableEntry = (uint)(g_paging_tables.shadowRam[offset - 1] << 16 | value);
+                pageTableEntry = ((uint32_t)g_paging_tables.shadowRam[offset - 1] << 16) | value;
             }
         }
         Log_Write(LOG_CAT_MMS, LOG_TRACE, "PT W A=%o PT=%d VPN=%d SEXI=%d V=%o => 0x%08X (%s)\n",  address, pageTable, pageTableAddress & 0x3F, STS_SEXI, value,  pageTableEntry, GetPageTableEntryDebugInfo(pageTableEntry));
@@ -220,7 +220,7 @@ uint GetPageTableEntry(uint pageTable, uint VPN,PageTableMode ptm)
 
     if (STS_SEXI)
     {
-        PTe = (uint)(g_paging_tables.shadowRam[pageTableAddress] << 16 | g_paging_tables.shadowRam[pageTableAddress + 1]);
+        PTe = ((uint32_t)g_paging_tables.shadowRam[pageTableAddress] << 16) | g_paging_tables.shadowRam[pageTableAddress + 1];
     }
     else
     {
@@ -262,7 +262,7 @@ uint GetPageTableEntryForDebugger(uint pageTable, uint VPN, PageTableMode ptm)
         uint offset = SHADOW_RAM_EXTENDED_MODE_16PT - g_paging_tables.shadowRamAddress;
         uint pageTableAddress = ((pageTable << 6) | VPN) << 1;
         pageTableAddress += offset;
-        PTe = (uint)(g_paging_tables.shadowRam[pageTableAddress] << 16 | g_paging_tables.shadowRam[pageTableAddress + 1]);
+        PTe = ((uint32_t)g_paging_tables.shadowRam[pageTableAddress] << 16) | g_paging_tables.shadowRam[pageTableAddress + 1];
     }
     else
     {
